+++ src/ska_mid_dish_manager/devices/DishManagerDS.py
@@ -2158,11 +2158,13 @@ class DishManager(SKAController):
     def ConfigureBand(self, json_string) -> DevVarLongStringArrayType:
         """Configure band according to JSON string supplied.
 
-        This command triggers the Dish to transition to the CONFIG Dish
-        Element Mode, and returns to the caller.
+                This command triggers the Dish to transition to the CONFIG Dish
+                Element Mode, re-applies the associated
+                band<n>pointingmodelparams, and returns to the caller.
+        .
 
-        :return: A tuple containing a return code and a string
-            message indicating status.
+                :return: A tuple containing a return code and a string
+                    message indicating status.
         """
         handler = self.get_command_object("ConfigureBand")
 
@@ -2181,8 +2183,10 @@ class DishManager(SKAController):
     )
     def ConfigureBand1(self, synchronise) -> DevVarLongStringArrayType:
         """This command triggers the Dish to transition to the CONFIG Dish
-        Element Mode, and returns to the caller. To configure the Dish to
-        operate in frequency band 1. On completion of the band
+        Element Mode, re-applies the associated band<n>pointingmodelparams,
+        and returns to the caller. To configure the Dish to
+        operate in frequency band 1 and re-apply the last updated
+        band1pointingmodelparams. On completion of the band
         configuration, Dish will automatically transition to Dish
         mode OPERATE.
 
@@ -2208,8 +2212,10 @@ class DishManager(SKAController):
         """Implemented as a Long Running Command.
 
         This command triggers the Dish to transition to the CONFIG Dish
-        Element Mode, and returns to the caller. To configure the Dish to
-        operate in frequency band 2. On completion of the band
+        Element Mode, re-applies the associated band<n>pointingmodelparams,
+        and returns to the caller. To configure the Dish to
+        operate in frequency band 2 and re-apply the last updated
+        band2pointingmodelparams. On completion of the band
         configuration, Dish will automatically transition to Dish
         mode OPERATE.
 
@@ -2233,8 +2239,10 @@ class DishManager(SKAController):
     )
     def ConfigureBand3(self, synchronise):  # pylint: disable=unused-argument
         """This command triggers the Dish to transition to the CONFIG Dish
-        Element Mode, and returns to the caller. To configure the Dish to
-        operate in frequency band 3. On completion of the band
+        Element Mode, re-applies the associated band<n>pointingmodelparams,
+        and returns to the caller. To configure the Dish to
+        operate in frequency band 3 and re-apply the last updated
+        band2pointingmodelparams. On completion of the band
         configuration, Dish will automatically transition to Dish
         mode OPERATE.
         """
@@ -2255,8 +2263,10 @@ class DishManager(SKAController):
     )
     def ConfigureBand4(self, synchronise):  # pylint: disable=unused-argument
         """This command triggers the Dish to transition to the CONFIG Dish
-        Element Mode, and returns to the caller. To configure the Dish to
-        operate in frequency band 4. On completion of the band
+        Element Mode, re-applies the associated band<n>pointingmodelparams,
+        and returns to the caller. To configure the Dish to
+        operate in frequency band 4 and re-apply the last updated
+        band4pointingmodelparams. On completion of the band
         configuration, Dish will automatically transition to Dish
         mode OPERATE.
         """
@@ -2277,8 +2287,10 @@ class DishManager(SKAController):
     )
     def ConfigureBand5a(self, synchronise):  # pylint: disable=unused-argument
         """This command triggers the Dish to transition to the CONFIG Dish
-        Element Mode, and returns to the caller. To configure the Dish to
-        operate in frequency band 5a. On completion of the band
+        Element Mode, re-applies the associated band<n>pointingmodelparams,
+        and returns to the caller. To configure the Dish to
+        operate in frequency band 5a and re-apply the last updated
+        band5apointingmodelparams. On completion of the band
         configuration, Dish will automatically transition to Dish
         mode OPERATE.
         """
@@ -2299,8 +2311,10 @@ class DishManager(SKAController):
     )
     def ConfigureBand5b(self, synchronise):  # pylint: disable=unused-argument
         """This command triggers the Dish to transition to the CONFIG Dish
-        Element Mode, and returns to the caller. To configure the Dish to
-        operate in frequency band 5b. On completion of the band
+        Element Mode, re-applies the associated band<n>pointingmodelparams,
+        and returns to the caller. To configure the Dish to
+        operate in frequency band 5b and re-apply the last updated
+        band5bpointingmodelparams. On completion of the band
         configuration, Dish will automatically transition to Dish
         mode OPERATE.
         """