+++ src/ska_mid_dish_manager/models/command_actions.py
@@ -755,6 +755,47 @@ class TrackStopAction(Action):
         )
 
 
+def apply_pointing_model(
+    band_param_name: str,
+    band_name: str,
+    task_callback,
+    logger: logging.Logger,
+    dish_manager_cm,
+):
+    """Apply pointing model parameters for a given band if they exist and are not all zeros.
+
+    Args:
+        band_param_name: The key in component_state for this band.
+        band_name: Name/identifier of the band (for logging).
+        task_callback: Callback to update task status on failure.
+        logger: Logger instance for info/debug/error messages.
+        dish_manager_cm: Component manager with component_state and update_pointing_model_params().
+
+    Returns:
+        TaskStatus.FAILED and error string if an error occurs, otherwise None.
+
+    """
+    try:
+        values = dish_manager_cm.component_state[band_param_name]
+
+        if values:
+            dish_manager_cm.update_pointing_model_params(band_param_name, values)
+            logger.debug(f"Pointing model for band {band_name} applied successfully")
+        else:
+            logger.debug(
+                f"Skipped applying pointing model for band {band_name} due to invalid params: []"
+            )
+
+    except (tango.DevFailed, ValueError) as err:
+        logger.error(f"Failed to apply pointing model for band {band_name}: {err}")
+        update_task_status(
+            task_callback,
+            status=TaskStatus.FAILED,
+            result=(ResultCode.FAILED, "Apply pointing model failed"),
+        )
+        return TaskStatus.FAILED, str(err)
+
+
 class ConfigureBandAction(Action):
     """Configure band on DS and SPFRx."""
 
@@ -931,7 +972,7 @@ class ConfigureBandActionSequence(Action):
         """Execute the defined action."""
         current_dish_mode = self.dish_manager_cm._component_state["dishmode"]
 
-        # Step 2: Operate mode action (final step)
+        # Step 3: Operate mode action (final step)
         operate_action = SetOperateModeAction(
             logger=self.logger,
             dish_manager_cm=self.dish_manager_cm,
@@ -943,7 +984,7 @@ class ConfigureBandActionSequence(Action):
 
         final_action = operate_action if current_dish_mode != DishMode.STOW else None
 
-        # Step 1: Configure band action
+        # Step 2: Configure band action
         configure_action = ConfigureBandAction(
             logger=self.logger,
             dish_manager_cm=self.dish_manager_cm,
@@ -956,7 +997,47 @@ class ConfigureBandActionSequence(Action):
             timeout_s=self.timeout_s,
         )
 
-        # Step 0: Pre-action if we need LP -> FP
+        # Step 0 :apply appropriate pointing models before configuring the band
+        # Case for Json arg configureband command
+        if self.data:
+            try:
+                data_json = json.loads(self.data)
+                dish_data = data_json.get("dish", {})
+                band_name = dish_data.get("receiver_band")
+                band_param_name = f"band{band_name}pointingmodelparams"
+
+                result = apply_pointing_model(
+                    band_param_name, band_name, task_callback, self.logger, self.dish_manager_cm
+                )
+                if result:
+                    return result
+
+            except (json.JSONDecodeError, ValueError) as err:
+                self.logger.error(f"Invalid JSON: {err}")
+                update_task_status(
+                    task_callback,
+                    status=TaskStatus.FAILED,
+                    result=(ResultCode.FAILED, "Invalid JSON in configureband command"),
+                )
+                return TaskStatus.FAILED, str(err)
+
+        else:
+            # Case for Non json arg configureband commands
+            if self.band in [Band.B5a, Band.B5b]:
+                enum_name = self.band.name
+                # Band name becomes '5a' or '5b'
+                band_name = enum_name[1:]
+            else:
+                band_name = str(self.band.value)
+            band_param_name = f"band{band_name}pointingmodelparams"
+
+            result = apply_pointing_model(
+                band_param_name, band_name, task_callback, self.logger, self.dish_manager_cm
+            )
+            if result:
+                return result
+
+        # Step 1: Pre-action if we need LP -> FP
         if current_dish_mode == DishMode.STANDBY_LP:
             pre_action = SetStandbyFPModeAction(
                 logger=self.logger,
@@ -966,7 +1047,6 @@ class ConfigureBandActionSequence(Action):
                 timeout_s=self.timeout_s,
             )
             return pre_action.execute(task_callback, task_abort_event, completed_response_msg)
-
         # If no LP -> FP transition is needed then start with ConfigureBand
         return configure_action.execute(task_callback, task_abort_event, completed_response_msg)