+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 ## unreleased
 *************
+- Updated ConfigureBand commands to automatically apply pointing models upon execution
 
 Version 9.3.0
 *************